+++ x/oracle/keeper/grpc_query.go
@@ -136,6 +136,31 @@ func (q querier) MissCounter(
 	}, nil
 }
 
+// SlashWindow queries the current slash window progress of the oracle.
+func (q querier) SlashWindow(
+	goCtx context.Context,
+	req *types.QuerySlashWindow,
+) (*types.QuerySlashWindowResponse, error) {
+	if req == nil {
+		return nil, status.Error(codes.InvalidArgument, "empty request")
+	}
+
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	params := q.GetParams(ctx)
+
+	slashWindow := params.SlashWindow
+	votePeriod := params.VotePeriod
+	currentBlock := uint64(ctx.BlockHeight())
+	votePeriodsPerSlashWindow := slashWindow / votePeriod
+
+	currentSlashWindow := currentBlock / votePeriodsPerSlashWindow
+	blocksIntoSlashWindow := currentBlock - (currentSlashWindow * slashWindow)
+
+	return &types.QuerySlashWindowResponse{
+		WindowProgress: blocksIntoSlashWindow / votePeriod,
+	}, nil
+}
+
 // AggregatePrevote queries an aggregate prevote of a validator.
 func (q querier) AggregatePrevote(
 	goCtx context.Context,